Correcting segfault in QApplication initialization
This bug occurs when exporting results to png. It is mayby related to the
fact that many qapplication are initialized in the same program (even if
they are correctly destroyed meanwhile).
Related bug report could be:
https://bugreports.qt.io/browse/QTBUG-35437

Solution was to instanciate Qt once at the program startup and to request
instance when needed.