correcting multiple phase bug (1st draft)
The label is updated for phases which field value is >= 0.5.
The problem is that, when using more than 2 phases, and when (at least) triple
contact point exists, the renormalization of the phases leads to a
neighbourhoud of this point where no phase have field >= 0.5.
The phase with default label (ie the background) was then polluting this zone.

The proposed solution (draft) is to assigned the label to the phase with maximum
field value. That implies to consider the background as a phase (with an
evolution equation ?).